+++ connect/src/wg/cosmo/platform/v1/platform_pb.ts
@@ -8,6 +8,7 @@
 import type { BinaryReadOptions, FieldList, JsonReadOptions, JsonValue, PartialMessage, PlainMessage } from "@bufbuild/protobuf";
 import { Message, proto3, protoInt64 } from "@bufbuild/protobuf";
 import { EnumStatusCode } from "../../common/common_pb.js";
+import { EventMeta } from "../../webhooks/events_pb.js";
 
 /**
  * @generated from enum wg.cosmo.platform.v1.AnalyticsViewGroupName
@@ -1195,42 +1196,47 @@ export class GetFederatedGraphsRequest extends Message<GetFederatedGraphsRequest
  */
 export class FederatedGraph extends Message<FederatedGraph> {
   /**
-   * @generated from field: string name = 1;
+   * @generated from field: string id = 1;
+   */
+  id = "";
+
+  /**
+   * @generated from field: string name = 2;
    */
   name = "";
 
   /**
-   * @generated from field: string routingURL = 2;
+   * @generated from field: string routingURL = 3;
    */
   routingURL = "";
 
   /**
-   * @generated from field: repeated string label_matchers = 3;
+   * @generated from field: repeated string label_matchers = 4;
    */
   labelMatchers: string[] = [];
 
   /**
-   * @generated from field: string lastUpdatedAt = 4;
+   * @generated from field: string lastUpdatedAt = 5;
    */
   lastUpdatedAt = "";
 
   /**
-   * @generated from field: bool isComposable = 5;
+   * @generated from field: bool isComposable = 6;
    */
   isComposable = false;
 
   /**
-   * @generated from field: string compositionErrors = 6;
+   * @generated from field: string compositionErrors = 7;
    */
   compositionErrors = "";
 
   /**
-   * @generated from field: int32 connectedSubgraphs = 7;
+   * @generated from field: int32 connectedSubgraphs = 8;
    */
   connectedSubgraphs = 0;
 
   /**
-   * @generated from field: repeated wg.cosmo.platform.v1.RequestSeriesItem requestSeries = 8;
+   * @generated from field: repeated wg.cosmo.platform.v1.RequestSeriesItem requestSeries = 9;
    */
   requestSeries: RequestSeriesItem[] = [];
 
@@ -1242,14 +1248,15 @@ export class FederatedGraph extends Message<FederatedGraph> {
   static readonly runtime: typeof proto3 = proto3;
   static readonly typeName = "wg.cosmo.platform.v1.FederatedGraph";
   static readonly fields: FieldList = proto3.util.newFieldList(() => [
-    { no: 1, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "routingURL", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 3, name: "label_matchers", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
-    { no: 4, name: "lastUpdatedAt", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 5, name: "isComposable", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
-    { no: 6, name: "compositionErrors", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 7, name: "connectedSubgraphs", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
-    { no: 8, name: "requestSeries", kind: "message", T: RequestSeriesItem, repeated: true },
+    { no: 1, name: "id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 2, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 3, name: "routingURL", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 4, name: "label_matchers", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+    { no: 5, name: "lastUpdatedAt", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 6, name: "isComposable", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
+    { no: 7, name: "compositionErrors", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 8, name: "connectedSubgraphs", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
+    { no: 9, name: "requestSeries", kind: "message", T: RequestSeriesItem, repeated: true },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FederatedGraph {
@@ -1515,6 +1522,11 @@ export class GetFederatedGraphByNameResponse extends Message<GetFederatedGraphBy
    */
   subgraphs: Subgraph[] = [];
 
+  /**
+   * @generated from field: string graphToken = 4;
+   */
+  graphToken = "";
+
   constructor(data?: PartialMessage<GetFederatedGraphByNameResponse>) {
     super();
     proto3.util.initPartial(data, this);
@@ -1526,6 +1538,7 @@ export class GetFederatedGraphByNameResponse extends Message<GetFederatedGraphBy
     { no: 1, name: "response", kind: "message", T: Response },
     { no: 2, name: "graph", kind: "message", T: FederatedGraph },
     { no: 3, name: "subgraphs", kind: "message", T: Subgraph, repeated: true },
+    { no: 4, name: "graphToken", kind: "scalar", T: 9 /* ScalarType.STRING */ },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetFederatedGraphByNameResponse {
@@ -2655,6 +2668,49 @@ export class Pagination extends Message<Pagination> {
 }
 
 /**
+ * @generated from message wg.cosmo.platform.v1.Sort
+ */
+export class Sort extends Message<Sort> {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id = "";
+
+  /**
+   * @generated from field: bool desc = 2;
+   */
+  desc = false;
+
+  constructor(data?: PartialMessage<Sort>) {
+    super();
+    proto3.util.initPartial(data, this);
+  }
+
+  static readonly runtime: typeof proto3 = proto3;
+  static readonly typeName = "wg.cosmo.platform.v1.Sort";
+  static readonly fields: FieldList = proto3.util.newFieldList(() => [
+    { no: 1, name: "id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 2, name: "desc", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
+  ]);
+
+  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Sort {
+    return new Sort().fromBinary(bytes, options);
+  }
+
+  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Sort {
+    return new Sort().fromJson(jsonValue, options);
+  }
+
+  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Sort {
+    return new Sort().fromJsonString(jsonString, options);
+  }
+
+  static equals(a: Sort | PlainMessage<Sort> | undefined, b: Sort | PlainMessage<Sort> | undefined): boolean {
+    return proto3.util.equals(Sort, a, b);
+  }
+}
+
+/**
  * @generated from message wg.cosmo.platform.v1.AnalyticsConfig
  */
 export class AnalyticsConfig extends Message<AnalyticsConfig> {
@@ -2673,6 +2729,11 @@ export class AnalyticsConfig extends Message<AnalyticsConfig> {
    */
   pagination?: Pagination;
 
+  /**
+   * @generated from field: optional wg.cosmo.platform.v1.Sort sort = 4;
+   */
+  sort?: Sort;
+
   constructor(data?: PartialMessage<AnalyticsConfig>) {
     super();
     proto3.util.initPartial(data, this);
@@ -2684,6 +2745,7 @@ export class AnalyticsConfig extends Message<AnalyticsConfig> {
     { no: 1, name: "date_range", kind: "message", T: DateRange },
     { no: 2, name: "filters", kind: "message", T: AnalyticsFilter, repeated: true },
     { no: 3, name: "pagination", kind: "message", T: Pagination },
+    { no: 4, name: "sort", kind: "message", T: Sort, opt: true },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): AnalyticsConfig {
@@ -4790,6 +4852,11 @@ export class CreateOrganizationWebhookConfigRequest extends Message<CreateOrgani
    */
   events: string[] = [];
 
+  /**
+   * @generated from field: repeated wg.cosmo.webhooks.EventMeta events_meta = 4;
+   */
+  eventsMeta: EventMeta[] = [];
+
   constructor(data?: PartialMessage<CreateOrganizationWebhookConfigRequest>) {
     super();
     proto3.util.initPartial(data, this);
@@ -4801,6 +4868,7 @@ export class CreateOrganizationWebhookConfigRequest extends Message<CreateOrgani
     { no: 1, name: "endpoint", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 2, name: "key", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 3, name: "events", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+    { no: 4, name: "events_meta", kind: "message", T: EventMeta, repeated: true },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): CreateOrganizationWebhookConfigRequest {
@@ -4981,6 +5049,86 @@ export class GetOrganizationWebhookConfigsResponse_Config extends Message<GetOrg
 }
 
 /**
+ * @generated from message wg.cosmo.platform.v1.GetOrganizationWebhookMetaRequest
+ */
+export class GetOrganizationWebhookMetaRequest extends Message<GetOrganizationWebhookMetaRequest> {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id = "";
+
+  constructor(data?: PartialMessage<GetOrganizationWebhookMetaRequest>) {
+    super();
+    proto3.util.initPartial(data, this);
+  }
+
+  static readonly runtime: typeof proto3 = proto3;
+  static readonly typeName = "wg.cosmo.platform.v1.GetOrganizationWebhookMetaRequest";
+  static readonly fields: FieldList = proto3.util.newFieldList(() => [
+    { no: 1, name: "id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+  ]);
+
+  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetOrganizationWebhookMetaRequest {
+    return new GetOrganizationWebhookMetaRequest().fromBinary(bytes, options);
+  }
+
+  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): GetOrganizationWebhookMetaRequest {
+    return new GetOrganizationWebhookMetaRequest().fromJson(jsonValue, options);
+  }
+
+  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): GetOrganizationWebhookMetaRequest {
+    return new GetOrganizationWebhookMetaRequest().fromJsonString(jsonString, options);
+  }
+
+  static equals(a: GetOrganizationWebhookMetaRequest | PlainMessage<GetOrganizationWebhookMetaRequest> | undefined, b: GetOrganizationWebhookMetaRequest | PlainMessage<GetOrganizationWebhookMetaRequest> | undefined): boolean {
+    return proto3.util.equals(GetOrganizationWebhookMetaRequest, a, b);
+  }
+}
+
+/**
+ * @generated from message wg.cosmo.platform.v1.GetOrganizationWebhookMetaResponse
+ */
+export class GetOrganizationWebhookMetaResponse extends Message<GetOrganizationWebhookMetaResponse> {
+  /**
+   * @generated from field: wg.cosmo.platform.v1.Response response = 1;
+   */
+  response?: Response;
+
+  /**
+   * @generated from field: repeated wg.cosmo.webhooks.EventMeta events_meta = 2;
+   */
+  eventsMeta: EventMeta[] = [];
+
+  constructor(data?: PartialMessage<GetOrganizationWebhookMetaResponse>) {
+    super();
+    proto3.util.initPartial(data, this);
+  }
+
+  static readonly runtime: typeof proto3 = proto3;
+  static readonly typeName = "wg.cosmo.platform.v1.GetOrganizationWebhookMetaResponse";
+  static readonly fields: FieldList = proto3.util.newFieldList(() => [
+    { no: 1, name: "response", kind: "message", T: Response },
+    { no: 2, name: "events_meta", kind: "message", T: EventMeta, repeated: true },
+  ]);
+
+  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetOrganizationWebhookMetaResponse {
+    return new GetOrganizationWebhookMetaResponse().fromBinary(bytes, options);
+  }
+
+  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): GetOrganizationWebhookMetaResponse {
+    return new GetOrganizationWebhookMetaResponse().fromJson(jsonValue, options);
+  }
+
+  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): GetOrganizationWebhookMetaResponse {
+    return new GetOrganizationWebhookMetaResponse().fromJsonString(jsonString, options);
+  }
+
+  static equals(a: GetOrganizationWebhookMetaResponse | PlainMessage<GetOrganizationWebhookMetaResponse> | undefined, b: GetOrganizationWebhookMetaResponse | PlainMessage<GetOrganizationWebhookMetaResponse> | undefined): boolean {
+    return proto3.util.equals(GetOrganizationWebhookMetaResponse, a, b);
+  }
+}
+
+/**
  * @generated from message wg.cosmo.platform.v1.UpdateOrganizationWebhookConfigRequest
  */
 export class UpdateOrganizationWebhookConfigRequest extends Message<UpdateOrganizationWebhookConfigRequest> {
@@ -5005,7 +5153,12 @@ export class UpdateOrganizationWebhookConfigRequest extends Message<UpdateOrgani
   events: string[] = [];
 
   /**
-   * @generated from field: bool shouldUpdateKey = 5;
+   * @generated from field: repeated wg.cosmo.webhooks.EventMeta events_meta = 5;
+   */
+  eventsMeta: EventMeta[] = [];
+
+  /**
+   * @generated from field: bool should_update_key = 6;
    */
   shouldUpdateKey = false;
 
@@ -5021,7 +5174,8 @@ export class UpdateOrganizationWebhookConfigRequest extends Message<UpdateOrgani
     { no: 2, name: "endpoint", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 3, name: "key", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 4, name: "events", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
-    { no: 5, name: "shouldUpdateKey", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
+    { no: 5, name: "events_meta", kind: "message", T: EventMeta, repeated: true },
+    { no: 6, name: "should_update_key", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): UpdateOrganizationWebhookConfigRequest {
@@ -5920,14 +6074,9 @@ export class GetMetricsErrorRateResponse extends Message<GetMetricsErrorRateResp
   response?: Response;
 
   /**
-   * @generated from field: repeated wg.cosmo.platform.v1.MetricsSeriesItem series = 2;
+   * @generated from field: repeated wg.cosmo.platform.v1.MetricsErrorRateSeriesItem series = 2;
    */
-  series: MetricsSeriesItem[] = [];
-
-  /**
-   * @generated from field: repeated wg.cosmo.platform.v1.MetricsSeriesItem errorSeries = 3;
-   */
-  errorSeries: MetricsSeriesItem[] = [];
+  series: MetricsErrorRateSeriesItem[] = [];
 
   constructor(data?: PartialMessage<GetMetricsErrorRateResponse>) {
     super();
@@ -5938,8 +6087,7 @@ export class GetMetricsErrorRateResponse extends Message<GetMetricsErrorRateResp
   static readonly typeName = "wg.cosmo.platform.v1.GetMetricsErrorRateResponse";
   static readonly fields: FieldList = proto3.util.newFieldList(() => [
     { no: 1, name: "response", kind: "message", T: Response },
-    { no: 2, name: "series", kind: "message", T: MetricsSeriesItem, repeated: true },
-    { no: 3, name: "errorSeries", kind: "message", T: MetricsSeriesItem, repeated: true },
+    { no: 2, name: "series", kind: "message", T: MetricsErrorRateSeriesItem, repeated: true },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetMetricsErrorRateResponse {
@@ -5956,6 +6104,55 @@ export class GetMetricsErrorRateResponse extends Message<GetMetricsErrorRateResp
 
   static equals(a: GetMetricsErrorRateResponse | PlainMessage<GetMetricsErrorRateResponse> | undefined, b: GetMetricsErrorRateResponse | PlainMessage<GetMetricsErrorRateResponse> | undefined): boolean {
     return proto3.util.equals(GetMetricsErrorRateResponse, a, b);
+  }
+}
+
+/**
+ * @generated from message wg.cosmo.platform.v1.MetricsErrorRateSeriesItem
+ */
+export class MetricsErrorRateSeriesItem extends Message<MetricsErrorRateSeriesItem> {
+  /**
+   * @generated from field: string timestamp = 1;
+   */
+  timestamp = "";
+
+  /**
+   * @generated from field: float requestRate = 2;
+   */
+  requestRate = 0;
+
+  /**
+   * @generated from field: float errorRate = 3;
+   */
+  errorRate = 0;
+
+  constructor(data?: PartialMessage<MetricsErrorRateSeriesItem>) {
+    super();
+    proto3.util.initPartial(data, this);
+  }
+
+  static readonly runtime: typeof proto3 = proto3;
+  static readonly typeName = "wg.cosmo.platform.v1.MetricsErrorRateSeriesItem";
+  static readonly fields: FieldList = proto3.util.newFieldList(() => [
+    { no: 1, name: "timestamp", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 2, name: "requestRate", kind: "scalar", T: 2 /* ScalarType.FLOAT */ },
+    { no: 3, name: "errorRate", kind: "scalar", T: 2 /* ScalarType.FLOAT */ },
+  ]);
+
+  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): MetricsErrorRateSeriesItem {
+    return new MetricsErrorRateSeriesItem().fromBinary(bytes, options);
+  }
+
+  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): MetricsErrorRateSeriesItem {
+    return new MetricsErrorRateSeriesItem().fromJson(jsonValue, options);
+  }
+
+  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): MetricsErrorRateSeriesItem {
+    return new MetricsErrorRateSeriesItem().fromJsonString(jsonString, options);
+  }
+
+  static equals(a: MetricsErrorRateSeriesItem | PlainMessage<MetricsErrorRateSeriesItem> | undefined, b: MetricsErrorRateSeriesItem | PlainMessage<MetricsErrorRateSeriesItem> | undefined): boolean {
+    return proto3.util.equals(MetricsErrorRateSeriesItem, a, b);
   }
 }