+++ controlplane/test/label.test.ts
@@ -22,7 +22,6 @@ import {
 } from '../src/core/test-util';
 import Keycloak from '../src/core/services/Keycloak';
 import { MockPlatformWebhookService } from '../src/core/webhooks/PlatformWebhookService';
-import { MockPrometheusClient } from '../src/core/prometheus/client';
 
 let dbname = '';
 
@@ -68,10 +67,6 @@ describe('Labels', (ctx) => {
 
     const platformWebhooks = new MockPlatformWebhookService();
 
-    const prometheus = new MockPrometheusClient({
-      apiUrl: 'http://localhost:9090',
-    });
-
     await server.register(fastifyConnectPlugin, {
       routes: routes({
         db: server.db,
@@ -81,7 +76,6 @@ describe('Labels', (ctx) => {
         keycloakRealm: realm,
         keycloakClient,
         platformWebhooks,
-        prometheus,
         webBaseUrl,
       }),
     });
@@ -184,10 +178,6 @@ describe('Labels', (ctx) => {
 
     const platformWebhooks = new MockPlatformWebhookService();
 
-    const prometheus = new MockPrometheusClient({
-      apiUrl: 'http://localhost:9090',
-    });
-
     await server.register(fastifyConnectPlugin, {
       routes: routes({
         db: server.db,
@@ -197,7 +187,6 @@ describe('Labels', (ctx) => {
         keycloakRealm: realm,
         keycloakClient,
         platformWebhooks,
-        prometheus,
         webBaseUrl,
       }),
     });
@@ -325,10 +314,6 @@ describe('Labels', (ctx) => {
 
     const platformWebhooks = new MockPlatformWebhookService();
 
-    const prometheus = new MockPrometheusClient({
-      apiUrl: 'http://localhost:9090',
-    });
-
     await server.register(fastifyConnectPlugin, {
       routes: routes({
         db: server.db,
@@ -338,7 +323,6 @@ describe('Labels', (ctx) => {
         keycloakRealm: realm,
         keycloakClient,
         platformWebhooks,
-        prometheus,
         webBaseUrl,
       }),
     });